requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
ansible==11.6.0
ansible-core==2.18.6
ansible-pylibssh==1.2.2
ansible-runner==2.4.1
arrow==1.3.0
async-lru==2.0.5
attrs==25.3.0
backports.ssl==0.0.9
backports.ssl_match_hostname==3.7.0.1
bcrypt==4.3.0
bidict==0.23.1
binaryornot==0.4.4
certifi==2025.4.26
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
ciscoconfparse==1.9.52
ciscoconfparse2==0.8.22
ciscoisesdk==2.3.1
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
coverage==7.8.2
cryptography==45.0.3
Deprecated==1.2.18
dict2xml==1.7.6
dill==0.4.0
distro==1.9.0
dnacentersdk==2.9.1
dnspython==2.7.0
et_xmlfile==2.0.0
f5-icontrol-rest==1.3.13
fastjsonschema==2.21.1
frozenlist==1.6.0
future==1.0.0
genie==25.4
genie.libs.clean==25.4
genie.libs.conf==25.4
genie.libs.filetransferutils==25.4
genie.libs.health==25.4
genie.libs.ops==25.4
genie.libs.parser==25.4
genie.libs.robot==25.4
genie.libs.sdk==25.4
genie.telemetry==25.4
genie.trafficgen==25.4
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.72.1
h11==0.16.0
hier-config==2.3.1
hypothesis==6.131.18
idna==3.10
iniconfig==2.1.0
IxNetwork==9.30.2212.8
ixnetwork_restpy==1.7.0
Jinja2==3.1.6
jsonpickle==4.1.1
junit-xml==1.9
libpass==1.9.1
lockfile==0.12.2
loguru==0.7.2
lxml==5.4.0
macaddress==2.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
ncclient==0.6.19
netaddr==1.3.0
netmiko==4.5.0
ntc_templates==7.9.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
paramiko==3.5.1
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pluggy==1.6.0
ply==3.11
prettytable==3.16.0
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pyasn1==0.6.0
pyasynchat==1.0.4
pyasyncore==1.0.4
pyats==25.4
pyats.aereport==25.4
pyats.aetest==25.4
pyats.async==25.4
pyats.connections==25.4
pyats.contrib==25.4
pyats.datastructures==25.4
pyats.easypy==25.4
pyats.kleenex==25.4
pyats.log==25.4
pyats.reporter==25.4
pyats.results==25.4
pyats.robot==25.4
pyats.tcl==25.4
pyats.topology==25.4
pyats.utils==25.4
pycparser==2.22
pyftpdlib==2.0.1
Pygments==2.19.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyserial==3.5
pysmi==1.6.2
pysnmp==6.1.4
pysnmpcrypto==0.0.4
pytest==8.4.0
pytest-cov==4.1.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.9.1
python-slugify==8.0.4
python-socketio==5.11.2
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
resolvelib==1.0.1
rest.connector==25.4
rich==14.0.0
robotframework==7.3
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
scp==0.15.0
scrypt==0.8.27
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sortedcontainers==2.4.0
text-unidecode==1.3
textfsm==1.1.3
tftpy==0.8.0
toml==0.10.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.4
types-python-dateutil==2.9.0.20250516
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.0
tzdata==2025.2
unicon==25.4
unicon.plugins==25.4
urllib3==2.4.0
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.2.3
xlwt==1.3.0
xmltodict==0.12.0
yamale==6.0.0
yamllint==1.37.1
yang.connector==25.4
yarl==1.20.0
